+++ src/types/Marker.tsx
@@ -1,0 +1,1 @@
+export type Marker = 'border' | 'checkmark' | 'fade' | 'none'