+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # NativeColorPicker Changelog
 
+## 1.1.0 beta 1 / 1.1.0 (2019-06-03)
+#### New Feature
+- added `animate` option
+
+#### Change
+- switched from `expo` to `expo-linear-gradient`
+
+#### Breaking Changes
+- dropped support for Expo < 33.0.0
+
 ## 1.0.0 beta 3.4 / 1.0.0 (2019-04-24)
 #### Documentation
 - added preview image to README.md